chore: make completion and misspelling chekc provider standalone
